+++ nsqlbproducer/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix(nsqlbproducer): correctly declare the v2
 * fix(nsqlbproducer): missing error wrapping
+* build(deps): update `github.com/Scalingo/go-utils/nsqproducer` to `v3`
 
 ## v2.0.0