+++ nsqlbproducer/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+## v2.0.1
+
 * fix(nsqlbproducer): correctly declare the v2
 * fix(nsqlbproducer): missing error wrapping
 * build(deps): update `github.com/Scalingo/go-utils/nsqproducer` to `v3`